+++ localstack/utils/http.py
